% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/category_match_indicator.R
\name{category_match_indicator}
\alias{category_match_indicator}
\title{category_match_indicator}
\usage{
category_match_indicator(categories_str, category, sep = ",")
}
\arguments{
\item{categories_str}{todo}

\item{category}{todo}
}
\value{
todo
}
\description{
category_match_indicator
}
\examples{
category_match_indicator("C, C++, R", "R")
}
