% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements_to_vector.R
\name{elements_to_vector}
\alias{elements_to_vector}
\title{elements_to_vector}
\usage{
elements_to_vector(
  my_string,
  sep = ",",
  remove_whitespaces = TRUE,
  unique = TRUE
)
}
\arguments{
\item{my_string}{descr}

\item{sep}{descr}

\item{remove_whitespaces}{descr}

\item{unique}{descr}
}
\value{
a vector
}
\description{
elements_to_vector
}
\examples{
elements_to_vector("R, Pyhton, C++, Java")
}
