% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encodeAsBinary.R
\name{encode_as_binary}
\alias{encode_as_binary}
\title{encodeAsBinary}
\usage{
encode_as_binary(df, categories_col, sep = ",", keep_categories_col = FALSE)
}
\arguments{
\item{df}{todo}

\item{categories_col}{todo}

\item{sep}{todo}

\item{keep_categories_col}{todo}
}
\value{
todo
}
\description{
encodeAsBinary
}
\examples{
df <- tibble::tibble(First_name=c("Marina", "Jane", "Elliot"),
Programming_languages=c("C, Scala","C, C++, R, Python", "C++, Python"))
df |>  encode_as_binary(Programming_languages)
}
