\name{ENCODER}
\alias{ENCODER}
\title{
ENCODER: off-target sequence reads for DNA copy number detection with whole-exome sequencing.
}
\description{
Generates DNA copy number profiles from whole exome sequencing bam files.
}
\usage{
ENCODER(bamFolder, destinationFolder, referenceFolder, whichControl, ncpu, captureRegionsBedFile)
}

\arguments{
  \item{bamFolder}{path to folder with the BAM files.}
  \item{destinationFolder}{path to folder in which output should be written.}
  \item{referenceFolder}{path to folder with the annotation files generated by \code{\link{preENCODER}} }
  \item{whichControl}{file that indicates the index of the reference BAM-file for every sample. For instance, if the file list is Ref1.bam; Tumor1.bam, and Ref1.bam (with index 1 in file list) should be used as a reference for both samples, c(1,1) or rep(1,2) would be appropriate.}
  \item{ncpu}{number of CPUs available for calculations.}
  \item{captureRegionsBedFile}{Optional; path to the capture design, should have a bed file format. If included, statistics on the overlap of peaks called by MACS and the capture regions will be provided.}
}
\details{
DNA copy number profiles directly from whole exome sequencing (WES) based on the captured genomic regions is complicated by the uneven distribution of exons throughout the genome and non-uniform sequence capture. We developed ENCODER (ENhanced COpy number Detection from Exome Reads) which eludes these problems by exploiting the off-target sequence reads.
ENCODER counts the reads outside the captured regions and generates a DNA copy number profile with an even distribution and uniform sequence depth over the whole genome. The number of datapoints depends on the number of total reads in the WES bam file. Generally, a 20kb bin size works for most WES samples. 
Before running ENCODER the function inputStructure should be used to generate the input files for ENCODER. 

}
\value{
  \item{log.txt}{log file of ENCODER.}
  \item{BamBaiMacsFiles}{folder with the files made by ENCODER needed to run the algorithm.}
  \item{CNAprofiles}{folder with the output generate by ENCODER}
}
\references{
ENCODER: DNA copy number detection from off-target sequence data. Thomas Kuilman, Arno Velds, Kristel Kemper, Marco Ranzani, Lorenzo Bombardelli, Guotai Xu, Julian de Ruiter, Bauke Ylstra, Jos Jonkers, Sven Rottenberg, Lodewyk F. Wessels, David J. Adams, Daniel S. Peeper, Oscar Krijgsman. Submitted for publication. 
}
\author{
Thomas Kuilman (t.kuilman@nki.nl)
Oscar Krijgsman (o.krijgsman@nki.nl)
}

\examples{
## Generate DNA copy number data from whole exome sequence data with annotation as generated by \code{\link{preENCODER}}.
  \dontrun{ENCODER("./PATH_TO_BAMFOLDER/", "./PATH_TO_OUTPUTFOLDER/", "./PATH_TO_FOLDER/hg19_20kb/", rep(2, 3), 3)} ## No captureRegionsBedFile
  \dontrun{ENCODER("./PATH_TO_BAMFOLDER/", "./PATH_TO_OUTPUTFOLDER/", "./PATH_TO_FOLDER/hg19_20kb/", rep(2, 3), 3, "./PATH_TO_CAPTUREREGIONSBEDFILE/captureRegionsBedFile.bed")} ## CaptureRegionsBedFile included in arguments
}
