% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_categories.R
\name{find_categories}
\alias{find_categories}
\title{find_categories}
\usage{
find_categories(df, categories_col, sep = ",")
}
\arguments{
\item{df}{a dtaframe}

\item{categories_col}{the name of the column containing string with categories inside}

\item{sep}{separator used in the strings. Ex : ,"}
}
\value{
a vector that list all categories in the considered column of the dataframe
}
\description{
find_categories
}
\examples{
df <- tibble::tibble(First_name=c("Marina", "Jane", "Elliot"),
Programming_languages=c("C, Scala","C, C++, R, Python", "C++, Python"))
df |> find_categories(Programming_languages)
}
