% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onehot_encoder.R
\name{onehot_encoder}
\alias{onehot_encoder}
\title{onehot_encoder}
\usage{
onehot_encoder(X_train, X_test = NULL, cat_columns)
}
\arguments{
\item{X_train}{X_train A `tibble` or `data.frame` representing the
training data set containing some categorical features/columns.}

\item{X_test}{X_test A `tibble` or `data.frame` representing the test set,
containing some set of categorical features/columns. Default = NULL.}

\item{cat_columns}{A character vector containing the names of the
categorical columns in the tibble that should be encoded.}
}
\value{
A list with processed training and test sets, in which the named
categorical columns are replaced with their encodings.
}
\description{
This function encodes categorical variables using the popular onehot method
for each category.
}
\examples{
onehot_encoder(
X_train = mtcars,
cat_columns = c("gear", "carb"))

}
