% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_function.R
\name{temp_function}
\alias{temp_function}
\title{This is a temporary function to create our repo.}
\usage{
temp_function()
}
\value{
numeric vector
}
\description{
This is a temporary function to create our repo.
}
\examples{
temp_function()
}
