% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/category_match_indicator.R
\name{category_match_indicator}
\alias{category_match_indicator}
\title{category_match_indicator}
\usage{
category_match_indicator(categories, category, sep = ",")
}
\arguments{
\item{categories}{a String. Ex : "Blue, Green, Yellow"}

\item{category}{a String. Ex : "Blue". The goal is to check if this category exists in categories = "Blue, Green, Yellow"}

\item{sep}{the separator used in the string. Ex : in categories = "Blue, Green, Yellow", sep=",".}
}
\value{
a Boolean. Is category in categories ? True or False.
}
\description{
category_match_indicator
}
\examples{
category_match_indicator("Blue, Green, Yellow", "Blue")
category_match_indicator("Blue, Green, Yellow", "Red")
category_match_indicator("Blue, Green, Yellow", "low")
category_match_indicator("C, C++, R", "R")
}
