% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_categories_in_string.R
\name{find_categories_in_string}
\alias{find_categories_in_string}
\title{find_categories_in_string}
\usage{
find_categories_in_string(my_string, sep = ",", unique = TRUE)
}
\arguments{
\item{my_string}{a string with separators. Ex : "A, B, C"}

\item{sep}{separator used in the string. Ex : ","}

\item{unique}{if TRUE, eliminates duplicates (keeps only one instance of each category)}
}
\value{
a vector that holds strings, with each string representing a category
}
\description{
find_categories_in_string
}
\examples{
find_categories_in_string("C++, C, Scala")
find_categories_in_string("C++; C; Scala", sep=";")
find_categories_in_string("C++, C, Scala, C")
find_categories_in_string("C++, C, Scala, C", unique=FALSE)
}
