% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integer_encoder.R
\name{integer_encoder}
\alias{integer_encoder}
\title{Integer Encoder}
\usage{
integer_encoder(data, feature, as_factor = TRUE)
}
\arguments{
\item{data}{Input data frame}

\item{feature}{Unquoted form of the feature/column to encode}

\item{as_factor}{Whether the factors should be converted to factors or pure integer}
}
\description{
Integer Encoder
}
\examples{
set.seed(11)
integer_encoder(iris[sample(1:150, 10),], Species)
}
