% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilabel_binarizer.R
\name{multi_label_binarizer}
\alias{multi_label_binarizer}
\alias{multi_label_binariser}
\title{Multi-label Binarizer}
\usage{
multi_label_binarizer(data, feature, sep = "#", as_integer = FALSE)

multi_label_binariser(data, feature, sep = "#", as_integer = FALSE)
}
\arguments{
\item{data}{Input data frame}

\item{feature}{Unquoted form of the feature/column to encode}

\item{sep}{Inner separator of the labels}

\item{as_integer}{Whether the boolean should be converted to integer or not}
}
\value{
Data frame with additional feature columns
}
\description{
Multi-label Binarizer
}
