% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_hot_encoder.R
\name{one_hot_encoder}
\alias{one_hot_encoder}
\title{One-hot Encoder}
\usage{
one_hot_encoder(data, feature, as_integer = FALSE)
}
\arguments{
\item{data}{Input data frame}

\item{feature}{Unquoted form of the feature/column to encode}

\item{as_integer}{Wether the boolean should be converted to integer or not}
}
\description{
One-hot Encoder
}
\examples{
set.seed(11)
one_hot_encoder(iris[sample(1:150, 10),], Species)
}
